src/detector/requirements.txt
kafka-python
scikit-learn
numpy
pandas
hdfs
requests
